+++ fe/src/components/Inquire/InquirePage.module.scss
@@ -1,5 +1,5 @@
-@use "../../../styles/variables.scss" as *;
-@use "../../../styles/mixins.scss" as *;
+@use "../../styles/variables.scss" as *;
+@use "../../styles/mixins.scss" as *;
 
 .container {
   position: relative;
